+++ nequip/data/AtomicData.py
@@ -33,7 +33,6 @@ _DEFAULT_LONG_FIELDS: Set[str] = {
 }
 _DEFAULT_NODE_FIELDS: Set[str] = {
     AtomicDataDict.POSITIONS_KEY,
-    AtomicDataDict.WEIGHTS_KEY,
     AtomicDataDict.NODE_FEATURES_KEY,
     AtomicDataDict.NODE_ATTRS_KEY,
     AtomicDataDict.ATOMIC_NUMBERS_KEY,
@@ -117,8 +116,16 @@ def _process_dict(kwargs, ignore_fields=[]):
             # Any property used as an index must be long (or byte or bool, but those are not relevant for atomic scale systems)
             # int32 would pass later checks, but is actually disallowed by torch
             kwargs[k] = torch.as_tensor(v, dtype=torch.long)
+        elif isinstance(v, bool):
+            kwargs[k] = torch.as_tensor(v)
         elif isinstance(v, np.ndarray):
             if np.issubdtype(v.dtype, np.floating):
+                kwargs[k] = torch.as_tensor(v, dtype=torch.get_default_dtype())
+            else:
+                kwargs[k] = torch.as_tensor(v)
+        elif isinstance(v, list):
+            ele_dtype = np.array(v).dtype
+            if np.issubdtype(ele_dtype, np.floating):
                 kwargs[k] = torch.as_tensor(v, dtype=torch.get_default_dtype())
             else:
                 kwargs[k] = torch.as_tensor(v)
@@ -199,7 +206,9 @@ class AtomicData(Data):
         **kwargs: other data, optional.
     """
 
-    def __init__(self, irreps: Dict[str, e3nn.o3.Irreps] = {}, **kwargs):
+    def __init__(
+        self, irreps: Dict[str, e3nn.o3.Irreps] = {}, _validate: bool = True, **kwargs
+    ):
 
         # empty init needed by get_example
         if len(kwargs) == 0 and len(irreps) == 0:
@@ -207,47 +216,49 @@ class AtomicData(Data):
             return
 
         # Check the keys
-        AtomicDataDict.validate_keys(kwargs)
-        _process_dict(kwargs)
+        if _validate:
+            AtomicDataDict.validate_keys(kwargs)
+            _process_dict(kwargs)
 
         super().__init__(num_nodes=len(kwargs["pos"]), **kwargs)
 
-        # Validate shapes
-        assert self.pos.dim() == 2 and self.pos.shape[1] == 3
-        assert self.edge_index.dim() == 2 and self.edge_index.shape[0] == 2
-        if "edge_cell_shift" in self and self.edge_cell_shift is not None:
-            assert self.edge_cell_shift.shape == (self.num_edges, 3)
-            assert self.edge_cell_shift.dtype == self.pos.dtype
-        if "cell" in self and self.cell is not None:
-            assert (self.cell.shape == (3, 3)) or (
-                self.cell.dim() == 3 and self.cell.shape[1:] == (3, 3)
-            )
-            assert self.cell.dtype == self.pos.dtype
-        if "node_features" in self and self.node_features is not None:
-            assert self.node_features.shape[0] == self.num_nodes
-            assert self.node_features.dtype == self.pos.dtype
-        if "node_attrs" in self and self.node_attrs is not None:
-            assert self.node_attrs.shape[0] == self.num_nodes
-            assert self.node_attrs.dtype == self.pos.dtype
-
-        if (
-            AtomicDataDict.ATOMIC_NUMBERS_KEY in self
-            and self.atomic_numbers is not None
-        ):
-            assert self.atomic_numbers.dtype in _TORCH_INTEGER_DTYPES
-        if "batch" in self and self.batch is not None:
-            assert self.batch.dim() == 2 and self.batch.shape[0] == self.num_nodes
-            # Check that there are the right number of cells
+        if _validate:
+            # Validate shapes
+            assert self.pos.dim() == 2 and self.pos.shape[1] == 3
+            assert self.edge_index.dim() == 2 and self.edge_index.shape[0] == 2
+            if "edge_cell_shift" in self and self.edge_cell_shift is not None:
+                assert self.edge_cell_shift.shape == (self.num_edges, 3)
+                assert self.edge_cell_shift.dtype == self.pos.dtype
             if "cell" in self and self.cell is not None:
-                cell = self.cell.view(-1, 3, 3)
-                assert cell.shape[0] == self.batch.max() + 1
+                assert (self.cell.shape == (3, 3)) or (
+                    self.cell.dim() == 3 and self.cell.shape[1:] == (3, 3)
+                )
+                assert self.cell.dtype == self.pos.dtype
+            if "node_features" in self and self.node_features is not None:
+                assert self.node_features.shape[0] == self.num_nodes
+                assert self.node_features.dtype == self.pos.dtype
+            if "node_attrs" in self and self.node_attrs is not None:
+                assert self.node_attrs.shape[0] == self.num_nodes
+                assert self.node_attrs.dtype == self.pos.dtype
 
-        # Validate irreps
-        # __*__ is the only way to hide from torch_geometric
-        self.__irreps__ = AtomicDataDict._fix_irreps_dict(irreps)
-        for field, irreps in self.__irreps__:
-            if irreps is not None:
-                assert self[field].shape[-1] == irreps.dim
+            if (
+                AtomicDataDict.ATOMIC_NUMBERS_KEY in self
+                and self.atomic_numbers is not None
+            ):
+                assert self.atomic_numbers.dtype in _TORCH_INTEGER_DTYPES
+            if "batch" in self and self.batch is not None:
+                assert self.batch.dim() == 2 and self.batch.shape[0] == self.num_nodes
+                # Check that there are the right number of cells
+                if "cell" in self and self.cell is not None:
+                    cell = self.cell.view(-1, 3, 3)
+                    assert cell.shape[0] == self.batch.max() + 1
+
+            # Validate irreps
+            # __*__ is the only way to hide from torch_geometric
+            self.__irreps__ = AtomicDataDict._fix_irreps_dict(irreps)
+            for field, irreps in self.__irreps__:
+                if irreps is not None:
+                    assert self[field].shape[-1] == irreps.dim
 
     @classmethod
     def from_points(
@@ -419,7 +430,11 @@ class AtomicData(Data):
             **add_fields,
         )
 
-    def to_ase(self, type_mapper=None) -> Union[List[ase.Atoms], ase.Atoms]:
+    def to_ase(
+        self,
+        type_mapper=None,
+        extra_fields: List[str] = [],
+    ) -> Union[List[ase.Atoms], ase.Atoms]:
         """Build a (list of) ``ase.Atoms`` object(s) from an ``AtomicData`` object.
 
         For each unique batch number provided in ``AtomicDataDict.BATCH_KEY``,
@@ -429,12 +444,18 @@ class AtomicData(Data):
         Args:
             type_mapper: if provided, will be used to map ``ATOM_TYPES`` back into
                 elements, if the configuration of the ``type_mapper`` allows.
+            extra_fields: fields other than those handled explicitly (currently
+                those defining the structure as well as energy, per-atom energy,
+                and forces) to include in the output object. Per-atom (per-node)
+                quantities will be included in ``arrays``; per-graph and per-edge
+                quantities will be included in ``info``.
 
         Returns:
             A list of ``ase.Atoms`` objects if ``AtomicDataDict.BATCH_KEY`` is in self
             and is not None. Otherwise, a single ``ase.Atoms`` object is returned.
         """
         positions = self.pos
+        edge_index = self[AtomicDataDict.EDGE_INDEX_KEY]
         if positions.device != torch.device("cpu"):
             raise TypeError(
                 "Explicitly move this `AtomicData` to CPU using `.to()` before calling `to_ase()`."
@@ -456,6 +477,21 @@ class AtomicData(Data):
         force = getattr(self, AtomicDataDict.FORCE_KEY, None)
         do_calc = energy is not None or force is not None
 
+        assert (
+            len(
+                set(extra_fields).intersection(
+                    [  # exclude those that are special for ASE and that we process seperately
+                        AtomicDataDict.POSITIONS_KEY,
+                        AtomicDataDict.CELL_KEY,
+                        AtomicDataDict.PBC_KEY,
+                        AtomicDataDict.ATOMIC_NUMBERS_KEY,
+                    ]
+                    + AtomicDataDict.ALL_ENERGY_KEYS
+                )
+            )
+            == 0
+        ), "Cannot specify typical keys as `extra_fields` for atoms output"
+
         if cell is not None:
             cell = cell.view(-1, 3, 3)
         if pbc is not None:
@@ -473,8 +509,11 @@ class AtomicData(Data):
             if batch is not None:
                 mask = batch == batch_idx
                 mask = mask.view(-1)
+                # if both ends of the edge are in the batch, the edge is in the batch
+                edge_mask = mask[edge_index[0]] & mask[edge_index[1]]
             else:
                 mask = slice(None)
+                edge_mask = slice(None)
 
             mol = ase.Atoms(
                 numbers=atomic_nums[mask].view(-1),  # must be flat for ASE
@@ -492,6 +531,26 @@ class AtomicData(Data):
                 if force is not None:
                     fields["forces"] = force[mask].cpu().numpy()
                 mol.calc = SinglePointCalculator(mol, **fields)
+
+            # add other information
+            for key in extra_fields:
+                if key in _NODE_FIELDS:
+                    # mask it
+                    mol.arrays[key] = (
+                        self[key][mask].cpu().numpy().reshape(mask.sum(), -1)
+                    )
+                elif key in _EDGE_FIELDS:
+                    mol.info[key] = (
+                        self[key][edge_mask].cpu().numpy().reshape(edge_mask.sum(), -1)
+                    )
+                elif key == AtomicDataDict.EDGE_INDEX_KEY:
+                    mol.info[key] = self[key][:, edge_mask].cpu().numpy()
+                elif key in _GRAPH_FIELDS:
+                    mol.info[key] = self[key][batch_idx].cpu().numpy().reshape(-1)
+                else:
+                    raise RuntimeError(
+                        f"Extra field `{key}` isn't registered as node/edge/graph"
+                    )
 
             batch_atoms.append(mol)
 
@@ -528,7 +587,8 @@ class AtomicData(Data):
 
     @classmethod
     def from_AtomicDataDict(cls, data: AtomicDataDict.Type):
-        return cls(**data)
+        # it's an AtomicDataDict, so don't validate-- assume valid:
+        return cls(_validate=False, **data)
 
     @property
     def irreps(self):