+++ nequip/nn/_atomwise.py
@@ -109,8 +109,8 @@ class PerSpeciesScaleShift(GraphModuleMixin, torch.nn.Module):
         self,
         field: str,
         num_types: int,
-        shifts: List[float],
-        scales: List[float],
+        shifts: Optional[List[float]],
+        scales: Optional[List[float]],
         arguments_in_dataset_units: bool,
         out_field: Optional[str] = None,
         scales_trainable: bool = False,
@@ -118,6 +118,7 @@ class PerSpeciesScaleShift(GraphModuleMixin, torch.nn.Module):
         irreps_in={},
     ):
         super().__init__()
+        self.num_types = num_types
         self.field = field
         self.out_field = f"shifted_{field}" if out_field is None else out_field
         self._init_irreps(
@@ -170,14 +171,19 @@ class PerSpeciesScaleShift(GraphModuleMixin, torch.nn.Module):
         return data
 
     def update_for_rescale(self, rescale_module):
+        if hasattr(rescale_module, "related_scale_keys"):
+            if self.out_field not in rescale_module.related_scale_keys:
+                return
         if self.arguments_in_dataset_units and rescale_module.has_scale:
             logging.debug(
                 f"PerSpeciesScaleShift's arguments were in dataset units; rescaling:\n"
-                f"Original scales {self.scales} shifts: {self.shifts}"
+                f"Original scales {self.scales if self.has_scales else 'n/a'} shifts: {self.shifts if self.has_shifts else 'n/a'}"
             )
             with torch.no_grad():
                 if self.has_scales:
                     self.scales.div_(rescale_module.scale_by)
                 if self.has_shifts:
                     self.shifts.div_(rescale_module.scale_by)
-            logging.debug(f"New scales {self.scales} shifts: {self.shifts}")
+            logging.debug(
+                f"New scales {self.scales if self.has_scales else 'n/a'} shifts: {self.shifts if self.has_shifts else 'n/a'}"
+            )