+++ nequip/scripts/train.py
@@ -1,6 +1,7 @@
 """ Train a network."""
 import logging
 import argparse
+import warnings
 
 # This is a weird hack to avoid Intel MKL issues on the cluster when this is called as a subprocess of a process that has itself initialized PyTorch.
 # Since numpy gets imported later anyway for dataset stuff, this shouldn't affect performance.
@@ -15,19 +16,19 @@ import e3nn
 import e3nn.util.jit
 
 from nequip.model import model_from_config
-from nequip.utils import Config
-from nequip.data import dataset_from_config
+from nequip.utils import Config, instantiate
+from nequip.data import dataset_from_config, register_fields
 from nequip.utils.test import assert_AtomicData_equivariant, set_irreps_debug
 from nequip.utils import load_file, dtype_from_name
-from nequip.scripts.logger import set_up_script_logger
+from nequip.scripts._logger import set_up_script_logger
 
 default_config = dict(
     root="./",
     run_name="NequIP",
     wandb=False,
     wandb_project="NequIP",
-    compile_model=False,
     model_builders=[
+        "SimpleIrrepsConfig",
         "EnergyModel",
         "PerSpeciesRescale",
         "ForceOutput",
@@ -129,6 +130,9 @@ def _set_global_options(config):
 
     e3nn.set_optimization_defaults(**config.get("e3nn_optimization_defaults", {}))
 
+    # Register fields:
+    instantiate(register_fields, all_args=config)
+
 
 def fresh_start(config):
 
@@ -176,9 +180,8 @@ def fresh_start(config):
 
     logging.info("Successfully built the network...")
 
-    if config.compile_model:
-        final_model = e3nn.util.jit.script(final_model)
-        logging.info("Successfully compiled model...")
+    # by doing this here we check also any keys custom builders may have added
+    _check_old_keys(config)
 
     # Equivar test
     if config.equivariance_test > 0:
@@ -267,6 +270,17 @@ def restart(config):
     trainer.set_dataset(dataset, validation_dataset)
 
     return trainer
+
+
+def _check_old_keys(config) -> None:
+    """check ``config`` for old/depricated keys and emit corresponding errors/warnings"""
+    # compile_model
+    k = "compile_model"
+    if k in config:
+        if config[k]:
+            raise ValueError("the `compile_model` option has been removed")
+        else:
+            warnings.warn("the `compile_model` option has been removed")
 
 
 if __name__ == "__main__":