+++ CHANGELOG.md
@@ -7,7 +7,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Most recent change on the bottom.
 
 
-## [Unreleased]
+## [Unreleased] 
+
+## [0.5.2] - 2022-02-04
+### Added
+- Model builders may now process only the configuration
+- Allow irreps to optionally be specified through the simplified keys `l_max`, `parity`, and `num_features`
+- `wandb.watch` via `wandb_watch` option
+- Allow polynomial cutoff _p_ values besides 6.0
+- `nequip-evaluate` now sets a default `r_max` taken from the model for the dataset config
+- Support multiple rescale layers in trainer
+- `AtomicData.to_ase` supports arbitrary fields
+- `nequip-evaluate` can now output arbitrary fields to an XYZ file
+- `nequip-evaluate` reports which frame in the original dataset was used as input for each output frame
+
+### Changed
+- `minimal.yaml`, `minimal_eng.yaml`, and `example.yaml` now use the simplified irreps options `l_max`, `parity`, and `num_features`
+- Default value for `resnet` is now `False`
+
+### Fixed
+- Handle one of `per_species_shifts`/`scales` being `null` when the other is a dataset statistc
+- `include_frames` now works with ASE datasets
+- no training data labels in input_data
+- Average number of neighbors no longer crashes sometimes when not all nodes have neighbors (small cutoffs)
+- Handle field registrations correctly in `nequip-evaluate`
+
+### Removed
+- `compile_model`
 
 ## [0.5.1] - 2022-01-13
 ### Added